old/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
matplotlib
tqdm
sentence-transformers
spacy
gensim

